+++ .openpublishing.redirection.json
@@ -944,6 +944,10 @@
             "redirect_url": "/dotnet/csharp"
         },
         {
+            "source_path": "docs/csharp/codedoc.md",
+            "redirect_url": "/dotnet/csharp/language-reference/xmldoc/recommended-tags"
+        },
+        {
             "source_path": "docs/csharp/deconstruct.md",
             "redirect_url": "/dotnet/csharp/fundamentals/functional/deconstruct"
         },
