+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/Program.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace xmldoc
+{
+    class Program
+    {
+        static void Main(string[] args)
+        {
+            Console.WriteLine("Hello World!");
+        }
+    }
+
+    
+}
