+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/idstrings.cs
@@ -1,0 +1,110 @@
+namespace N
+{
+    /// <summary>
+    /// Enter description here for class X.
+    /// ID string generated is "T:N.X".
+    /// </summary>
+    public unsafe class X
+    {
+        /// <summary>
+        /// Enter description here for the first constructor.
+        /// ID string generated is "M:N.X.#ctor".
+        /// </summary>
+        public X() { }
+
+        /// <summary>
+        /// Enter description here for the second constructor.
+        /// ID string generated is "M:N.X.#ctor(System.Int32)".
+        /// </summary>
+        /// <param name="i">Describe parameter.</param>
+        public X(int i) { }
+
+        /// <summary>
+        /// Enter description here for field q.
+        /// ID string generated is "F:N.X.q".
+        /// </summary>
+        public string q;
+
+        /// <summary>
+        /// Enter description for constant PI.
+        /// ID string generated is "F:N.X.PI".
+        /// </summary>
+        public const double PI = 3.14;
+
+        /// <summary>
+        /// Enter description for method f.
+        /// ID string generated is "M:N.X.f".
+        /// </summary>
+        /// <returns>Describe return value.</returns>
+        public int f() { return 1; }
+
+        /// <summary>
+        /// Enter description for method bb.
+        /// ID string generated is "M:N.X.bb(System.String,System.Int32@,System.Void*)".
+        /// </summary>
+        /// <param name="s">Describe parameter.</param>
+        /// <param name="y">Describe parameter.</param>
+        /// <param name="z">Describe parameter.</param>
+        /// <returns>Describe return value.</returns>
+        public int bb(string s, ref int y, void* z) { return 1; }
+
+        /// <summary>
+        /// Enter description for method gg.
+        /// ID string generated is "M:N.X.gg(System.Int16[],System.Int32[0:,0:])".
+        /// </summary>
+        /// <param name="array1">Describe parameter.</param>
+        /// <param name="array">Describe parameter.</param>
+        /// <returns>Describe return value.</returns>
+        public int gg(short[] array1, int[,] array) { return 0; }
+
+        /// <summary>
+        /// Enter description for operator.
+        /// ID string generated is "M:N.X.op_Addition(N.X,N.X)".
+        /// </summary>
+        /// <param name="x">Describe parameter.</param>
+        /// <param name="xx">Describe parameter.</param>
+        /// <returns>Describe return value.</returns>
+        public static X operator +(X x, X xx) { return x; }
+
+        /// <summary>
+        /// Enter description for property.
+        /// ID string generated is "P:N.X.prop".
+        /// </summary>
+        public int prop { get { return 1; } set { } }
+
+        /// <summary>
+        /// Enter description for event.
+        /// ID string generated is "E:N.X.d".
+        /// </summary>
+        public event D d;
+
+        /// <summary>
+        /// Enter description for property.
+        /// ID string generated is "P:N.X.Item(System.String)".
+        /// </summary>
+        /// <param name="s">Describe parameter.</param>
+        /// <returns></returns>
+        public int this[string s] { get { return 1; } }
+
+        /// <summary>
+        /// Enter description for class Nested.
+        /// ID string generated is "T:N.X.Nested".
+        /// </summary>
+        public class Nested { }
+
+        /// <summary>
+        /// Enter description for delegate.
+        /// ID string generated is "T:N.X.D".
+        /// </summary>
+        /// <param name="i">Describe parameter.</param>
+        public delegate void D(int i);
+
+        /// <summary>
+        /// Enter description for operator.
+        /// ID string generated is "M:N.X.op_Explicit(N.X)~System.Int32".
+        /// </summary>
+        /// <param name="x">Describe parameter.</param>
+        /// <returns>Describe return value.</returns>
+        public static explicit operator int(X x) { return 1; }
+    }
+}
