+++ docs/csharp/language-reference/xmldoc/snippets/xmldoc/xmldoc.xml
@@ -1,0 +1,115 @@
+<?xml version="1.0"?>
+<doc>
+    <assembly>
+        <name>xmldoc</name>
+    </assembly>
+    <members>
+        <member name="T:N.X">
+            <summary>
+            Enter description here for class X.
+            ID string generated is "T:N.X".
+            </summary>
+        </member>
+        <member name="M:N.X.#ctor">
+            <summary>
+            Enter description here for the first constructor.
+            ID string generated is "M:N.X.#ctor".
+            </summary>
+        </member>
+        <member name="M:N.X.#ctor(System.Int32)">
+            <summary>
+            Enter description here for the second constructor.
+            ID string generated is "M:N.X.#ctor(System.Int32)".
+            </summary>
+            <param name="i">Describe parameter.</param>
+        </member>
+        <member name="F:N.X.q">
+            <summary>
+            Enter description here for field q.
+            ID string generated is "F:N.X.q".
+            </summary>
+        </member>
+        <member name="F:N.X.PI">
+            <summary>
+            Enter description for constant PI.
+            ID string generated is "F:N.X.PI".
+            </summary>
+        </member>
+        <member name="M:N.X.f">
+            <summary>
+            Enter description for method f.
+            ID string generated is "M:N.X.f".
+            </summary>
+            <returns>Describe return value.</returns>
+        </member>
+        <member name="M:N.X.bb(System.String,System.Int32@,System.Void*)">
+            <summary>
+            Enter description for method bb.
+            ID string generated is "M:N.X.bb(System.String,System.Int32@,System.Void*)".
+            </summary>
+            <param name="s">Describe parameter.</param>
+            <param name="y">Describe parameter.</param>
+            <param name="z">Describe parameter.</param>
+            <returns>Describe return value.</returns>
+        </member>
+        <member name="M:N.X.gg(System.Int16[],System.Int32[0:,0:])">
+            <summary>
+            Enter description for method gg.
+            ID string generated is "M:N.X.gg(System.Int16[],System.Int32[0:,0:])".
+            </summary>
+            <param name="array1">Describe parameter.</param>
+            <param name="array">Describe parameter.</param>
+            <returns>Describe return value.</returns>
+        </member>
+        <member name="M:N.X.op_Addition(N.X,N.X)">
+            <summary>
+            Enter description for operator.
+            ID string generated is "M:N.X.op_Addition(N.X,N.X)".
+            </summary>
+            <param name="x">Describe parameter.</param>
+            <param name="xx">Describe parameter.</param>
+            <returns>Describe return value.</returns>
+        </member>
+        <member name="P:N.X.prop">
+            <summary>
+            Enter description for property.
+            ID string generated is "P:N.X.prop".
+            </summary>
+        </member>
+        <member name="E:N.X.d">
+            <summary>
+            Enter description for event.
+            ID string generated is "E:N.X.d".
+            </summary>
+        </member>
+        <member name="P:N.X.Item(System.String)">
+            <summary>
+            Enter description for property.
+            ID string generated is "P:N.X.Item(System.String)".
+            </summary>
+            <param name="s">Describe parameter.</param>
+            <returns></returns>
+        </member>
+        <member name="T:N.X.Nested">
+            <summary>
+            Enter description for class Nested.
+            ID string generated is "T:N.X.Nested".
+            </summary>
+        </member>
+        <member name="T:N.X.D">
+            <summary>
+            Enter description for delegate.
+            ID string generated is "T:N.X.D".
+            </summary>
+            <param name="i">Describe parameter.</param>
+        </member>
+        <member name="M:N.X.op_Explicit(N.X)~System.Int32">
+            <summary>
+            Enter description for operator.
+            ID string generated is "M:N.X.op_Explicit(N.X)~System.Int32".
+            </summary>
+            <param name="x">Describe parameter.</param>
+            <returns>Describe return value.</returns>
+        </member>
+    </members>
+</doc>
