+++ docs/csharp/toc.yml
@@ -296,8 +296,6 @@ items:
       href: expression-trees-summary.md
   - name: Native interoperability
     href: programming-guide/interop/index.md
-  - name: Documenting your code
-    href: codedoc.md
   - name: Versioning
     href: versioning.md
 - name: How-to C# articles
