--- samples/snippets/csharp/concepts/codedoc/xml-comment.cs
@@ -1,6 +1,0 @@
-﻿/// <summary>
-/// This class does something.
-/// </summary>
-public class SomeClass
-{
-}
